requirements.txt
# Telegram Bot
python-telegram-bot==20.7
qrcode[pil]==7.4.2
pillow==10.1.0

# Voice Processing
SpeechRecognition==3.10.0
pyaudio==0.2.14
openai==1.3.0
pydub==0.25.1

# Integrations
spotipy==2.23.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# OAuth Server
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Hardware
gpiozero==2.0.1
RPi.GPIO==0.7.1

# Security & Utils
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
